+++ apps/desktop/package.json
@@ -18,6 +18,7 @@
     "chart.js": "^2.7.2",
     "electron-webpack-ts": "^2.0.2",
     "f1-laps-js-bridge": "file:../../modules/js-bridge",
+    "hyperapp": "^1.2.8",
     "lodash": "^4.17.10",
     "moment": "^2.22.2",
     "typescript": "^3.0.1"
