+++ apps/desktop/tsconfig.json
@@ -1,7 +1,8 @@
 {
-    "extends": "./node_modules/electron-webpack/tsconfig-base.json",
-
-    "compilerOptions": {
-        "strict": false
-    }
+  "extends": "./node_modules/electron-webpack/tsconfig-base.json",
+  "compilerOptions": {
+    "jsx": "react",
+    "jsxFactory": "h",
+    "strict": false
+  }
 }
